+++ src/install-apt.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
+
+echo "## Installing APT tools ##"
 apt update
 apt-get -y install software-properties-common