+++ webapp/index.php
@@ -1,0 +1,8 @@
+<?php
+
+use Fhooe\WebDockDashboard\Dashboard;
+
+require_once "dashboard/vendor/autoload.php";
+
+$dashboard = new Dashboard();
+$dashboard->display();